- Fixing delete;

+++ src/main/java/org/coldis/library/persistence/keyvalue/KeyValueService.java
@@ -48,7 +48,10 @@ public class KeyValueService {
 	@Transactional(propagation = Propagation.REQUIRED)
 	public void delete(
 			final String key) {
-		this.repository.deleteById(key);
+		this.repository.flush();
+		if (this.repository.existsById(key)) {
+			this.repository.deleteById(key);
+		}
 	}
 
 	/**
